KHexEditor: Possible to load big files (>4GB) when compiled with 64-bit compiler.

+++ source/keditcommon.pas
@@ -345,7 +345,7 @@ type
     ErrorReason: TKEditSearchError;
     Options: TKEditSearchOptions;
     SelStart,
-    SelEnd: Integer;
+    SelEnd: Int64;
     TextToFind,
     TextToReplace: string;
   end;
